Move classes to subpackages
