+++ static-analysis-plugin/plugin/src/main/groovy/com/novoda/staticanalysis/StaticAnalysisPlugin.groovy
@@ -1,6 +1,9 @@
 package com.novoda.staticanalysis
 
-import com.novoda.staticanalysis.findbugs.FindbugsConfigurator
+import com.novoda.staticanalysis.internal.checkstyle.CheckstyleConfigurator
+import com.novoda.staticanalysis.internal.findbugs.FindbugsConfigurator
+import com.novoda.staticanalysis.internal.Violations
+import com.novoda.staticanalysis.internal.pmd.PmdConfigurator
 import org.gradle.api.NamedDomainObjectContainer
 import org.gradle.api.Plugin
 import org.gradle.api.Project
