+++ static-analysis-plugin/plugin/src/main/groovy/com/novoda/staticanalysis/internal/pmd/PmdViolationsEvaluator.groovy
@@ -1,4 +1,4 @@
-package com.novoda.staticanalysis
+package com.novoda.staticanalysis.internal.pmd
 
 import groovy.util.slurpersupport.GPathResult
